Move old index ID util fns to old index util module
- was in the shared util module
