+++ src/search/search-index-old/pipeline.js
@@ -2,7 +2,8 @@ import normalizeUrl from 'normalize-url'
 
 import transformPageText from 'src/util/transform-page-text'
 import { convertMetaDocId } from 'src/activity-logger'
-import { keyGen, DEFAULT_TERM_SEPARATOR, extractContent } from '../util'
+import { DEFAULT_TERM_SEPARATOR, extractContent } from '../util'
+import { keyGen } from './util'
 
 // Simply extracts the timestamp component out the ID of a visit or bookmark doc,
 //  which is the only data we want at the moment.
