+++ src/search/search-index-old/suggest.js
@@ -1,5 +1,5 @@
 import index from './'
-import { keyGen, removeKeyType } from '../util'
+import { keyGen, removeKeyType } from './util'
 
 /**
  * @param {string} [query=''] Plaintext query string to match against start of domain and tag names.
