+++ src/search/search-index-old/util.js
@@ -1,7 +1,21 @@
 import promiseLimit from 'promise-limit'
 
-import { removeKeyType } from '../util'
 import index, { indexQueue } from './'
+
+// Key generation functions
+export const keyGen = {
+    domain: key => `domain/${key}`,
+    tag: key => `tag/${key}`,
+    url: key => `url/${key}`,
+    term: key => `term/${key}`,
+    title: key => `title/${key}`,
+    visit: key => `visit/${key}`,
+    bookmark: key => `bookmark/${key}`,
+    _: key => key,
+}
+
+export const removeKeyType = key =>
+    key.replace(/^(term|title|visit|url|domain|tag|bookmark)\//, '')
 
 /**
  * @param {(args: any) => Promise<any>} fn Any async function to place on the index operations queue.
