+++ src/search/util.js
@@ -1,20 +1,7 @@
+import { keyGen } from './search-index-old/util'
+
 export const DEFAULT_TERM_SEPARATOR = /[|' .,\-|(\n)]+/
 export const URL_SEPARATOR = /[/?#=+& _.,\-|(\n)]+/
-
-export const removeKeyType = key =>
-    key.replace(/^(term|title|visit|url|domain|tag|bookmark)\//, '')
-
-// Key generation functions
-export const keyGen = {
-    domain: key => `domain/${key}`,
-    tag: key => `tag/${key}`,
-    url: key => `url/${key}`,
-    term: key => `term/${key}`,
-    title: key => `title/${key}`,
-    visit: key => `visit/${key}`,
-    bookmark: key => `bookmark/${key}`,
-    _: key => key,
-}
 
 /**
  * Handles splitting up searchable content into indexable terms. Terms are all
